Add Specta for TypeScript bindings generation

+++ src-tauri/src/data_manager/provider.rs
@@ -5,7 +5,7 @@ use std::sync::Arc;
 
 #[async_trait]
 pub trait DataProvider {
-    async fn get_next_entry(&mut self, number_plates: usize) -> Result<Arc<ColumnEntry>>;
+    async fn get_next_entry(&mut self, number_plates: i32) -> Result<Arc<ColumnEntry>>;
     fn skip(&mut self, count: i64) -> Result<()>;
     fn reset(&mut self) -> Result<()>;
     fn get_current_index(&self) -> usize;
